test: 단위 테스트 Mockito 적용

+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -4,7 +4,9 @@ import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import roomescape.global.error.exception.BadRequestException;
 import roomescape.global.error.exception.ConflictException;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.theme.dto.request.ThemeCreateRequest;
 import roomescape.theme.dto.response.ThemeCreateResponse;
 import roomescape.theme.dto.response.ThemeReadResponse;
@@ -16,6 +18,7 @@ import roomescape.theme.repository.ThemeRepository;
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
 
     public ThemeCreateResponse createTheme(ThemeCreateRequest request) {
         Theme newTheme = request.toEntity();
@@ -43,6 +46,9 @@ public class ThemeService {
     }
 
     public void deleteTheme(Long id) {
+        if (reservationRepository.existsByThemeId(id)) {
+            throw new BadRequestException("해당 테마에 예약된 내역이 존재하므로 삭제할 수 없습니다.");
+        }
         themeRepository.deleteById(id);
     }
 }
